add report; basic runtime
