+++ python/heterocl/devices.py
@@ -285,7 +285,7 @@ class platform(with_metaclass(env, object)):
           self.tool = tool(compile, *option_table[compile]) 
         
         if mode: # check tool mode 
-          modes = ["sw_sim", "hw_sim", "hw_exe", "debug"]
+          modes = ["sw_sim", "sw_exe", "hw_sim", "hw_exe", "debug"]
           assert mode in modes, \
               "supported tool mode: " + str(modes)
           self.tool.mode = mode
