+++ tests/test_runtime_build.py
@@ -46,32 +46,36 @@ def test_vivado_hls():
     if os.system("which vivado_hls >> /dev/null") != 0:
         return 
 
-    hcl.init()
-    A = hcl.placeholder((10, 32), "A")
-    def kernel(A):
-        B = hcl.compute(A.shape, lambda *args : A[args] + 1, "B")
-        C = hcl.compute(A.shape, lambda *args : B[args] + 1, "C")
-        D = hcl.compute(A.shape, lambda *args : C[args] * 2, "D")
-        return D
-    
-    target = hcl.platform.aws_f1
-    s = hcl.create_schedule([A], kernel)
-    s.to(kernel.B, target.xcel)
-    s.to(kernel.C, target.host)
-    target.config(compile="vivado_hls", mode="sw_sim")
-    f = hcl.build(s, target)
+    def test_sim(mode):
+        hcl.init()
+        A = hcl.placeholder((10, 32), "A")
+        def kernel(A):
+            B = hcl.compute(A.shape, lambda *args : A[args] + 1, "B")
+            C = hcl.compute(A.shape, lambda *args : B[args] + 1, "C")
+            D = hcl.compute(A.shape, lambda *args : C[args] * 2, "D")
+            return D
+        
+        target = hcl.platform.aws_f1
+        s = hcl.create_schedule([A], kernel)
+        s.to(kernel.B, target.xcel)
+        s.to(kernel.C, target.host)
+        target.config(compile="vivado_hls", mode=mode)
+        f = hcl.build(s, target)
 
-    np_A = np.random.randint(10, size=(10,32))
-    np_B = np.zeros((10,32))
+        np_A = np.random.randint(10, size=(10,32))
+        np_B = np.zeros((10,32))
 
-    hcl_A = hcl.asarray(np_A)
-    hcl_B = hcl.asarray(np_B, dtype=hcl.Int(32))
-    f(hcl_A, hcl_B)
-    ret_B = hcl_B.asnumpy()
+        hcl_A = hcl.asarray(np_A)
+        hcl_B = hcl.asarray(np_B, dtype=hcl.Int(32))
+        f(hcl_A, hcl_B)
+        ret_B = hcl_B.asnumpy()
 
-    for i in range(0, 10):
-      for j in range(0, 32):
-        assert ret_B[i, j] == (np_A[i, j] + 2) *2
+        for i in range(0, 10):
+          for j in range(0, 32):
+            assert ret_B[i, j] == (np_A[i, j] + 2) *2
+
+    test_sim("sw_sim")
+    test_sim("sw_exe")
 
 def test_mixed_stream():
     if os.system("which vivado_hls >> /dev/null") != 0:
@@ -170,7 +174,7 @@ def test_vitis():
 
         assert np.array_equal(hcl_D.asnumpy(), np_A + np_B + 1)
 
-    # test_arith_ops()
+    test_arith_ops()
     test_xrt_stream()
 
 
@@ -242,6 +246,6 @@ if __name__ == '__main__':
     test_debug_mode()
     test_vivado_hls()
     test_mixed_stream()
-    test_vitis()
+    # test_vitis()
     test_xilinx_sdsoc()
     test_intel_aocl()
