+++ tvm/src/template/vivado/run.tcl
@@ -10,7 +10,7 @@ set hls_prj out.prj
 open_project ${hls_prj} -reset
 
 # Top function of the design is "top"
-set_top top
+set_top test
 
 # Add design and testbench files
 add_files kernel.cpp
